pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "superskills"
version = "2.5.0"
description = "AI-powered automation skills for freelance coaching & training businesses"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Steff Vanhaverbeke", email = "steff@cs-workx.be"}
]
keywords = ["ai", "automation", "claude", "skills", "coaching", "training"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "google-genai>=0.2.0",
    "elevenlabs>=0.2.0",
    "pydub>=0.25.0",
    "audioop-lts>=0.2.0; python_version >= '3.13'",
    "Pillow>=10.0.0",
    "crawl4ai>=0.1.0",
    "beautifulsoup4>=4.12.0",
    "pyyaml>=6.0.0",
    "watchdog>=3.0.0",
    "rich>=13.0.0",
    "jsonschema>=4.0.0",
]

[project.scripts]
superskills = "cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
media = [
    "pydub>=0.25.0",
    "Pillow>=10.0.0",
]
all = [
    "superskills[dev,media]",
]

[project.urls]
Homepage = "https://github.com/CoachSteff/superskills"
Documentation = "https://github.com/CoachSteff/superskills#readme"
Repository = "https://github.com/CoachSteff/superskills"
Issues = "https://github.com/CoachSteff/superskills/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["superskills*", "cli*"]
exclude = ["tests*", "docs*", "archive*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.coverage.run]
source = ["superskills"]
omit = ["*/tests/*", "*/archive/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
